Add new dataset with hospitalizations per municipality
