+++ data_collection.sh
@@ -1,4 +1,4 @@
-python download_rivm_corona.py
+python download_rivm_map_data.py
 python merge_data.py
 python render_datasets.py
 python press_releases.py
