+++ render_datasets_old.py
@@ -8,6 +8,18 @@ import pandas
 import numpy
 
 
+def daily_data(data):
+
+    result = data.groupby('Datum')[['Aantal']].sum()
+
+    # manual edits
+    result.loc['2020-03-02', 'Aantal'] = 18
+
+    result['Aantal'] = result['Aantal'].astype(numpy.int64)
+    print(result.sort_index())
+    return result.sort_index()
+
+
 def wide_data(data):
 
     key_cols = [col for col in list(data) if col != 'Aantal']
@@ -32,13 +44,12 @@ def wide_data(data):
     df_mun.iloc[:, 3:] = df_mun.iloc[:, 3:].fillna(0).astype(int)
     return df_mun
 
-
 if __name__ == '__main__':
 
-    # df = pandas.read_csv(Path("data", "rivm_NL_covid19_hosp_municipality.csv"))
+    df = pandas.read_csv(Path("data", "rivm_corona_in_nl.csv"))
+    daily_data(df).to_csv(Path("data", "rivm_corona_in_nl_daily.csv"))
 
-    # df_wide = wide_data(df).to_csv(
-    #     Path("data", "rivm_corona_in_nl_table.csv"),
-    #     index=False
-    # )
-    pass
+    df_wide = wide_data(df).to_csv(
+        Path("data", "rivm_corona_in_nl_table.csv"),
+        index=False
+    )
